Converting Randomization Implementation
This commit takes the randomization out of stack_init and places it in pthread_attr_init instead. A boolean (or int for C code) is used as an argument to trigger the generation of a random value which is masked with 0x3fffff000000 to ensure it remains under the 0x400000000000 upper bound as defined in mmu.cc. Additionally a check bit is used, 0x300000000000, to ensure the bits in the farthest left position are set, and a new random number generated otherwise. This is to account for random values that do not have the lower bits set and would result in an unacceptably small value. For instance, the value 0x8872cde468b9, the left most 8 will not have any of its lower bits set, so when the mask is applied there will be no value in that place.